requirements.txt
# PDF Processing
pdfplumber==0.10.3
pytesseract==0.3.10

# ML/NLP (for future) - Mac compatible versions
transformers==4.36.0
torch==2.6.0
scikit-learn==1.3.2
scipy==1.11.4

# LLM
anthropic==0.18.1

# Data
pandas==2.1.4
numpy==1.26.2

# Utils
python-dotenv==1.0.0
tqdm==4.66.1

# API
fastapi==0.108.0
uvicorn==0.25.0

# Testing
pytest==7.4.3

#anthropic
#requests
#dotenv